+++ packages/engine/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.1.0
 
+- FIX: Added the needed styles to the body tag to avoid issues with the Flutter app.
 - Added new attestationToken option
 
 ## v1.0.0